Feature: Support `run` and `run-result` for strings
This does not have parity with the version in JVM NetLogo, since it
doesn't support accessing:

  * Local variables
  * Parameters
  * Agent context-specific primitives (e.g. `color`, `hatch`)

Disables a bunch of language test freebies that are no longer needed.

It's a little gross, but there are essentially three different ways
`run*` can work.  First, if the browser is providing a config for us to
use, we grab that (which will call out to the browser compiler,
populating the necessary values from the information in the UI).  If
that's not there, we're either in a `TortoiseFixture` or a
`DockingFixture` (assumedly), and call the JVM compiler.  In a
`DockingFixture`, we grab the necessary data out of the JVM workspace.
In a `TortoiseFixture`, the necessary data floats locally in the
`TortoiseFixture` object.

This is... not awful, but unpleasant.  Ideally, there shouldn't be three
different paths to the same data.  I feel like a nice, unified solution
to this might be to store the model code in the workspace somewhere.
But that sounds pretty cumbersome on the output JS, and I have fears
that it would be a pain to keep that state up-to-date.

This is all something to think about, going forward, though.  This
fragmented solution is another obstacle in the way of running the tests
in Node.