WIP/Needs Cleanup: Provide `nobody` from compiler, not engine
The final stuff in this commit (where the compiler is loaded into
Nashorn) needs cleanup.  This is a really crappy way to be loading the
compiler.  The compiler should be put on the classpath by the build, and
it should be loaded up in *all* of the JS engines, not just Nashorn.